Reorganization and cleanup before proceeding to further coding / reimplmentation
